requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
attrs==22.2.0
bcrypt==4.0.1
blessed==1.20.0
botocore==1.29.81
Brotli==1.0.9
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
contourpy==1.0.7
cryptography==39.0.2
cssselect2==0.7.0
cycler==0.11.0
Django==3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-simple-history==3.2.0
djangorestframework==3.14.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
fonttools==4.39.2
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
jmespath==1.0.1
jsonschema==3.2.0
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
matplotlib==3.7.1
numpy==1.24.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
paramiko==3.1.0
pathspec==0.10.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.5.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyNaCl==1.5.0
pyparsing==3.0.9
pypdf==3.7.0
pyphen==0.14.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
qrcode==7.4.2
reportlab==3.6.12
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
svglib==1.5.1
termcolor==1.1.0
texttable==1.6.7
tinycss2==1.2.1
typing_extensions==4.4.0
tzdata==2023.2
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
wcwidth==0.1.9
weasyprint==58.1
webencodings==0.5.1
websocket-client==0.59.0
xhtml2pdf==0.2.9
zipp==3.11.0
zopfli==0.2.2
